segmentation/MGL/requirements.txt
apex
cached-property>=1.5.2
cycler>=0.10.0
decorator>=4.4.2
future>=0.18.2
imageio>=2.9.0
kiwisolver>=1.3.1
matplotlib>=3.3.3
networkx>=2.5
pandas>=1.1.5
progress>=1.5
protobuf>=3.14.0
pyparsing>=2.4.7
python-dateutil>=2.8.1
pytz>=2021.1
pywavelets>=1.1.1
pyyaml>=5.3.1
scikit-image>=0.17.2
scipy>=1.5.4
tensorboardx>=2.1
thop>=0.0.31-2005241907
tifffile>=2020.9.3
torchstat>=0.0.7
torchsummary>=1.5.1
tqdm>=4.56.0
install
h5py>=3.1.0


